Fix TypeScript inconsistencies in question subsystem
- Move NewQuestionEventData and QuestionAnsweredEventData to types.ts
  alongside approval event data types
- Replace inline import() types with locally imported Question type
  in DaemonClient interface
- Add try-catch to answerQuestion in http-client.ts to match
  sendDecision error handling pattern
- Update QuestionContent handleSubmit/handleDecline to check
  result.success before proceeding
- Update Layout.tsx to import question event types from types.ts
